added welcome message to HTML
