+++ css/bingocard.css
@@ -1,0 +1,68 @@
+body{
+	margin-left: 1%;
+	margin-right: 1%;
+	font-family:Verdana, Geneva, Tahoma, sans-serif;
+	font-style: italic;
+	font-size: small;
+}
+
+ul {
+	list-style-type: none;
+}
+
+li {
+	width: 19%;
+}
+
+.flexContainer {
+	display:flex;
+	flex-wrap: wrap;
+	text-align: center;
+	display:none;
+}
+
+#filler {
+	flex: 8%;
+	background-color: black;
+}
+
+.bingoHeading {
+	flex:92%;
+	font-weight:bolder;
+	font-size: 2em;
+	background-color: firebrick;
+}
+
+.bingoHeading li {
+	float:left;
+}
+
+.pelotonHeading {
+	background-color: firebrick;
+	width: 8%;
+	font-weight:bolder;
+	font-size: 2em;
+}
+
+.pelotonHeading ul{
+	margin-top: 2em;
+}
+
+.pelotonHeading li {
+	height: 80px;
+	margin-top: 2%;
+}
+
+.squares {
+	width: 92%;
+}
+
+.squares li {
+	float: left;
+	border-style:ridge;
+	border-color:black;
+	border-width:thin;
+	padding-top: 1em;
+	height: 80px;
+	overflow-wrap: break-word;
+}
