+++ css/welcome.css
@@ -1,0 +1,27 @@
+.welcome  {
+    background-image: url('../images/cil-map.jpg');
+    opacity: .75;
+    background-repeat: no-repeat;
+    background-size: cover;
+}
+
+.welcomeContainer {
+    display: flex;
+    height: 10em;
+    justify-content: center;
+    padding-top: 5%;
+    padding-bottom: 40%;
+    font-size:x-large;
+    font-weight:bolder;
+}
+
+.welcomeText {
+    border-style: solid;
+    background-color:mediumaquamarine;
+    padding: 2%;
+}
+
+button {
+    padding: 2%;
+    margin-left: 45%;
+}
