+++ index.html
@@ -1,15 +1,29 @@
 <!DOCTYPE HTML>
 <html lang="en">
 	<head>
-		<link rel="stylesheet" type="text/css" href="bingocard.css">
-		<script type="module" src="card.js"></script>
+		<link rel="stylesheet" type="text/css" href="css/welcome.css">
+		<link rel="stylesheet" type="text/css" href="css/bingocard.css">
+		<script type="module" src="script/card.js"></script>
 		<meta charset="utf-8">
 		<meta name="description" content="Bingo card">
 		<title>Peloton Bingo</title>
 	</head>
 
-	<body id='displayCard'>	
+	<body>	
 		<section>
+			<section class="welcome"> 
+				<div class="welcomeContainer">
+					<div class="welcomeText">
+					<h1>
+						Welcome to the Central Illinois Bingo Card Generator!
+					</h1>
+					<button>Generate Card</button>
+					</div>
+
+				</div>
+
+
+			</section>
 			<div class="flexContainer">
 				<div id="filler">	
 				</div>
@@ -74,10 +88,6 @@
 					</ul>
 				</div>
 			</div>
-			
-
 		</section>	
-		
-	</body>
-	
+	</body>	
 </html>
